enable mounting with davfs2 (https://savannah.nongnu.org/projects/davfs2)
OPTIONS / was returning a "DAV: 2" header.  This causes davfs2 to reject the
attempt:

/sbin/mount.davfs: mounting failed; the server does not support WebDAV

Setting ignore_dav_header 1 in /etc/davfs2/davfs2.conf allows the mount to
continue, but this shouldn't be necessary.  Advertising Class 1 capability
in addition to Class 2 fixes this problem, while not affecting compatibility
with other clients (was still able to mount the device from Windows 10, for
instance).